备份

+++ tp5/application/gjcf/controller/Logout.php
@@ -9,11 +9,16 @@ use app\gjcf\api\Helper as HelperApi;
 
 class Logout extends Controller{
     public function Logout(){
-        HelperApi::TestLoginAndStatus($this);
+        $status = HelperApi::TestLoginAndStatus($this);
+        if(true !== $status){
+            return $status;
+        }
 
 
         Session::delete('userid');
-        $this->success('已退出', 'gjcf/login/index', 0, 1);
+        $json_arr = ['code' => 0, 'msg' => '已退出'];
+        return json_encode($json_arr);
+//        $this->success('已退出', 'gjcf/login/index', 0, 1);
     }
 }
